+++ src/components/drop_down_table.jsx
@@ -1,9 +1,10 @@
-import { Fragment, forwardRef, useMemo } from 'react';
+import { Fragment, forwardRef } from 'react';
 import PropTypes from 'prop-types';
 import { DropDown } from './drop_down';
 import { renderListBox } from './list_box_table/render_list_box';
 import { renderGroupLabel } from './list_box_table/render_group_label';
 import { renderOption } from './list_box_table/render_option';
+import { useNormalisedColumns } from '../hooks/use_normalised_columns';
 
 function renderNothing() {
   return null;
@@ -59,18 +60,7 @@ export const DropDownTable = forwardRef(
     },
     ref,
   ) => {
-    const columns = useMemo(
-      () =>
-        rawColumns.map((column) => {
-          if (typeof column === 'string') {
-            return {
-              name: column,
-            };
-          }
-          return column;
-        }),
-      [rawColumns],
-    );
+    const columns = useNormalisedColumns(rawColumns);
 
     return (
       <DropDown