+++ src/components/list_box_table/render_group_label.jsx
@@ -8,6 +8,10 @@ export function renderGroupLabel(props, componentState, componentProps) {
     renderTableGroupHeaderCell,
   } = componentProps;
 
+  const {
+    group: { key },
+  } = componentState;
+
   return renderTableGroupRow(
     {
       className: makeBEMClass(classPrefix, 'table-group-row'),
@@ -16,6 +20,7 @@ export function renderGroupLabel(props, componentState, componentProps) {
           ...props,
           colSpan: Object.keys(columns).length,
           className: makeBEMClass(classPrefix, 'table-group-header'),
+          id: key,
         },
         componentState,
         componentProps,